+++ Form/Screens/root.py
@@ -1,9 +1,7 @@
 from tkinter import *
 from VIStk.Objects import Root
-from VIStk.Structures.project import Project
 
 root = Root()
-project = Project()
 
 def do_focus(event=None):
     x,y = root.winfo_pointerxy()